Rework api directory structure and rename Status to ResponseStatus
