--- src/api/Project/index.js
@@ -1,8 +1,0 @@
-// @flow
-import createProject from './mutations/createProject'
-
-export default {
-  Mutation: {
-    createProject
-  }
-}
