+++ src/api/projects/index.js
@@ -1,8 +1,10 @@
 // @flow
+import createProject from './mutations/createProject'
 import createProjectUser from './mutations/createProjectUser'
 
 export default {
   Mutation: {
+    createProject,
     createProjectUser
   }
 }
