+++ src/api/schema/index.js
@@ -1,12 +1,11 @@
 // @flow
 import { makeExecutableSchema } from 'graphql-tools'
-import UserAPI from '@api/User'
-import ProjectAPI from '@api/Project'
-import ProjectUserAPI from '@api/ProjectUser'
+import usersAPI from '@api/users'
+import projectsAPI from '@api/projects'
 import Root from './types/Root.graphql'
 import Scalars from './scalars/scalars.graphql'
 import ScalarsResolvers from './scalars/resolvers'
-import Status from './types/Status.graphql'
+import ResponseStatus from './types/ResponseStatus.graphql'
 import ValidationError from './types/ValidationError.graphql'
 import User from './types/User.graphql'
 import Project from './types/Project.graphql'
@@ -15,13 +14,13 @@ import ProjectUser from './types/ProjectUser.graphql'
 const typeDefs = [
   Root,
   Scalars,
-  Status,
+  ResponseStatus,
   ValidationError,
   User,
   Project,
   ProjectUser
 ]
-const resolvers = [ScalarsResolvers, UserAPI, ProjectAPI, ProjectUserAPI]
+const resolvers = [ScalarsResolvers, usersAPI, projectsAPI]
 const schema = makeExecutableSchema({
   typeDefs,
   resolvers
